Moved utility functionality into separate files.
