+++ src/components/instructor/SetDetails/HomeworkSet.vue
@@ -43,7 +43,7 @@ import { cloneDeep } from 'lodash-es';
 
 import DateTimeInput from 'src/components/common/DateTimeInput.vue';
 import { HomeworkSet } from 'src/store/models';
-import { newHomeworkSet, copyProblemSet } from 'src/store/common';
+import { newHomeworkSet, copyProblemSet } from 'src/store/utils/problem_sets';
 import { useStore } from 'src/store';
 
 export default defineComponent({
