+++ src/store/utils/course_settings.ts
@@ -1,0 +1,10 @@
+// this has common utility functions for CourseSettings
+
+import { CourseSetting } from '../models';
+
+export function newCourseSetting(): CourseSetting {
+	return {
+		var: '',
+		value: ''
+	};
+}
